backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
opencv-python==4.10.0.84
torch==2.4.1
torchvision==0.19.1
easyocr==1.7.1
openai-whisper==20240930
spacy==3.7.6
mtcnn==0.1.1
insightface==0.7.3
numpy==1.26.4
python-dotenv==1.0.1
transformers==4.40.0
# NOTE: ffmpeg must be installed separately on the host system (not a pip package).
# Windows: winget install ffmpeg   |   Linux: apt install ffmpeg
